# backend/openapi.yaml
openapi: 3.0.3
info:
  title: JARS Cannabis Mobile App API
  description: API for the JARS cannabis mobile application
  version: 1.0.0
  contact:
    name: JARS API Support
    url: https://jars.app/support
    email: api-support@jars.app
  license:
    name: Proprietary
    url: https://jars.app/terms
servers:
  - url: https://api.jars.app/api/v1
    description: Production server
  - url: https://staging-api.jars.app/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        description:
          type: string
        imageUrl:
          type: string
          format: uri
        featured:
          type: boolean
        thcContent:
          type: number
          format: float
        cbdContent:
          type: number
          format: float
    CartItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        quantity:
          type: integer
          minimum: 1
        variant:
          type: string
        product:
          $ref: '#/components/schemas/Product'
...
    Order:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, confirmed, preparing, delivering, delivered, cancelled]
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        total:
          type: number
          format: float
        deliveryMethod:
          type: string
          enum: [delivery, pickup]
        deliveryAddress:
          type: object
        estimatedDelivery:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    Store:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        phone:
          type: string
        hours:
          type: object
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: List products with optional filters
      security: []
      parameters:
        - name: category
          in: query
          description: Filter by product category
          schema:
            type: string
        - name: search
          in: query
          description: Search products by name
          schema:
            type: string
        - name: minPrice
          in: query
          description: Minimum price filter
          schema:
            type: number
            format: float
        - name: maxPrice
          in: query
          description: Maximum price filter
          schema:
            type: number
            format: float
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: '#/components/schemas/Product'
                  relatedProducts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /orders:
    get:
      tags:
        - Orders
      summary: List user orders
      parameters:
        - name: status
          in: query
          description: Filter orders by status
          schema:
            type: string
            enum: [pending, confirmed, preparing, delivering, delivered, cancelled]
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: List of user orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Orders
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - deliveryMethod
              properties:
                deliveryMethod:
                  type: string
                  enum: [delivery, pickup]
                deliveryAddress:
                  type: object
                  properties:
                    street:
                      type: string
                    city:
                      type: string
                    state:
                      type: string
                    zipCode:
                      type: string
                paymentMethod:
                  type: string
                  enum: [card, cash]
                paymentDetails:
                  type: object
                tip:
                  type: number
                  format: float
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /stores/{id}:
    get:
      tags:
        - Stores
      summary: Get store details
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Store ID
          schema:
            type: string
      responses:
        '200':
          description: Store details
          content:
            application/json:
              schema:
                type: object
                properties:
                  store:
                    $ref: '#/components/schemas/Store'
        '404':
          description: Store not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...